[refactor] 결제 구조 변경 및 repo 분리
