+++ src/main/java/kr/hhplus/be/server/domain/payment/infra/PaymentJpaRepository.java
@@ -1,11 +1,12 @@
-package kr.hhplus.be.server.domain.payment.repository;
+package kr.hhplus.be.server.domain.payment.infra;
 
-import kr.hhplus.be.server.domain.payment.entity.Payment;
+import kr.hhplus.be.server.domain.payment.domain.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.sql.Timestamp;
 import java.util.List;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentJpaRepository extends JpaRepository<Payment, Long> {
+
     List<Payment> findOrderIdByStatusAndPaidAtAfter(String status, Timestamp days);
 }
